🚧 wip: lots of refactoring
