+++ clios/application.py
@@ -4,13 +4,13 @@ from typing import Annotated, Any, Callable
 import click
 from rich import print
 
-from clios.operator.operator import RootOperator
+from clios.core.operator import RootOperator
 
 from .cli import print_detail, print_list
 from .cli.ast_builder import ASTBuilder
 from .cli.tokenizer import tokenize
-from .operator.param_info import Input
-from .operator.utils import Implicit, get_operator_fn
+from .core.param_info import Input
+from .core.utils import Implicit, get_operator_fn
 from .registry import OperatorRegistry
