+++ tests/core/test_Parameter_init_execute_phase_validator.py
@@ -4,8 +4,8 @@ from typing import Annotated, Any
 import pytest
 from pydantic import BeforeValidator, Strict
 
-from clios.operator.operator_fn import Parameter, ParameterKind
-from clios.operator.param_info import Param
+from clios.core.operator_fn import Parameter, ParameterKind
+from clios.core.param_info import Param
 
 
 @pytest.fixture
@@ -32,49 +32,53 @@ def type_adapter_mock(mocker):
 
 def test_with_build_phase_strict(parameter, type_adapter_mock):
     param_type = Param(
-        build_phase_validators=[int, str], core_validation_phase="build", strict=True
+        execute_phase_validators=[int, str], core_validation_phase="build", strict=True
     )
     parameter_mock = parameter(param_type)
-    parameter_mock._init_build_phase_validator()
+    parameter_mock._init_execute_phase_validator()
 
     type_adapter_mock.assert_called_once_with(
-        Annotated[int, BeforeValidator(int), BeforeValidator(str), Strict()]
+        Annotated[Any, BeforeValidator(int), BeforeValidator(str)]
     )
-    assert parameter_mock._build_phase_validator == "TypeAdapter"
+    assert parameter_mock._execute_phase_validator == "TypeAdapter"
 
 
 def test_with_build_phase_nostrict(parameter, type_adapter_mock):
-    param_type = Param(build_phase_validators=[int, str], core_validation_phase="build")
+    param_type = Param(
+        execute_phase_validators=[int, str], core_validation_phase="build"
+    )
     parameter_mock = parameter(param_type)
-    parameter_mock._init_build_phase_validator()
+    parameter_mock._init_execute_phase_validator()
 
     type_adapter_mock.assert_called_once_with(
-        Annotated[int, BeforeValidator(int), BeforeValidator(str)]
+        Annotated[Any, BeforeValidator(int), BeforeValidator(str)]
     )
-    assert parameter_mock._build_phase_validator == "TypeAdapter"
+    assert parameter_mock._execute_phase_validator == "TypeAdapter"
 
 
 def test_with_execute_phase_nostrict(parameter, type_adapter_mock):
     param_type = Param(
-        build_phase_validators=[int, str], core_validation_phase="execute"
+        execute_phase_validators=[int, str], core_validation_phase="execute"
     )
     parameter_mock = parameter(param_type)
-    parameter_mock._init_build_phase_validator()
+    parameter_mock._init_execute_phase_validator()
 
     type_adapter_mock.assert_called_once_with(
-        Annotated[Any, BeforeValidator(int), BeforeValidator(str)]
+        Annotated[int, BeforeValidator(int), BeforeValidator(str)]
     )
-    assert parameter_mock._build_phase_validator == "TypeAdapter"
+    assert parameter_mock._execute_phase_validator == "TypeAdapter"
 
 
 def test_with_execute_phase_strict(parameter, type_adapter_mock):
     param_type = Param(
-        build_phase_validators=[int, str], core_validation_phase="execute", strict=True
+        execute_phase_validators=[int, str],
+        core_validation_phase="execute",
+        strict=True,
     )
     parameter_mock = parameter(param_type)
-    parameter_mock._init_build_phase_validator()
+    parameter_mock._init_execute_phase_validator()
 
     type_adapter_mock.assert_called_once_with(
-        Annotated[Any, BeforeValidator(int), BeforeValidator(str)]
+        Annotated[int, BeforeValidator(int), BeforeValidator(str), Strict()]
     )
-    assert parameter_mock._build_phase_validator == "TypeAdapter"
+    assert parameter_mock._execute_phase_validator == "TypeAdapter"
