+++ tests/core/test_get_typed_signature.py
@@ -1,6 +1,6 @@
 import inspect
 
-from clios.operator.utils import get_typed_signature
+from clios.core.utils import get_typed_signature
 
 
 def test_():
